# Tude-Web-App-master/backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Étude+ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
        role:
          type: string
          enum: [student, professor, admin]
        fullName:
          type: string
        profilePhoto:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        createdAt:
          type: string
        studentProfile:
          $ref: "#/components/schemas/StudentProfile"
          nullable: true
        professorProfile:
          $ref: "#/components/schemas/Professor"
          nullable: true
      required: [id, email, role, fullName, createdAt]
    StudentProfile:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        gradeLevel:
          type: string
          nullable: true
        schoolName:
          type: string
          nullable: true
        preferredSubjects:
          type: array
          items:
            type: string
        parentContact:
          type: string
          nullable: true
      required: [id, userId, preferredSubjects]
    Professor:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        fullName:
          type: string
        profilePhoto:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        subjects:
          type: array
          items:
            type: string
        gradeLevels:
          type: array
          items:
            type: string
        yearsOfExperience:
          type: number
          nullable: true
        bio:
          type: string
          nullable: true
        qualifications:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        rating:
          type: number
          nullable: true
        totalReviews:
          type: number
        totalStudents:
          type: number
        isVerified:
          type: boolean
        createdAt:
          type: string
      required: [id, userId, fullName, subjects, gradeLevels, status, totalReviews, totalStudents, isVerified, createdAt]
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, professor]
        fullName:
          type: string
        city:
          type: string
          nullable: true
      required: [email, password, role, fullName]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
...